Bugfix: MET in some events is a NaN. Such events are skipped now
The problem is extremely rare: a single event found among all the processed dataset. The nominal value of MET is not affected; the problem was spotted in a variation of unclustered MET, but other systematical variation have not been check consistently.

When such a problematic event is rejected, a warning is printed in the standard output.